src/race_core/requirements.txt
pyserial
paho-mqtt
click
